requirements.txt
Flask
speedtest-cli